tree view and websocket fix
