+++ build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 group = "com.software.codetime"
-version = "2.8.19"
+version = "2.8.20"
 
 repositories {
     mavenCentral()
