+++ src/main/java/com/software/codetime/toolwindows/WebviewCommandHandler.java
@@ -64,7 +64,7 @@ public class WebviewCommandHandler {
                 break;
             case "viewDashboard":
                 ApplicationManager.getApplication().invokeLater(() -> {
-                    UtilManager.launchUrl(ConfigManager.app_url + "/dashboard/code_time?view=summary");
+                    UtilManager.launchUrl(ConfigManager.app_url + "/dashboard/code_time");
                 });
                 break;
             case "softwareKpmDashboard":
